Fix namespaces in Mangos.World
