+++ Source/Services/Mangos.World/Program.cs
@@ -8,6 +8,7 @@ using Mangos.Configuration;
 using Mangos.Configuration.Store;
 using Mangos.Configuration.Xml;
 using Mangos.World.AI;
+using Mangos.World.AntiCheat;
 using Mangos.World.Auction;
 using Mangos.World.Battlegrounds;
 using Mangos.World.DataStores;
@@ -27,7 +28,7 @@ using Microsoft.VisualBasic.CompilerServices;
 
 namespace Mangos.World
 {
-	[StandardModule]
+    [StandardModule]
 	public sealed class Program
 	{
 		[Serializable]
