--- Source/Services/Mangos.World/ViolationType.cs
@@ -1,10 +1,0 @@
-namespace Mangos.World
-{
-	public enum ViolationType
-	{
-		AC_VIOLATION_NONE = 0,
-		AC_VIOLATION_SPEEDHACK_TIME = 1,
-		AC_VIOLATION_SPEEDHACK_MEM = 3,
-		AC_VIOLATION_MOVEMENT_Z = 2
-	}
-}
